Move chunked-vectors read-only code into a dedicated module
`chunked_vectors` kept the read-only view (`ChunkedVectorsRead`), its
`LiveReload` impl and the writer in flat sibling files. Reshape it to the
layout the other components use: a `read_only/` submodule split into
`mod.rs` / `lifecycle.rs` / `read_ops.rs` / `live_reload.rs`, with the
writer's impls split into `lifecycle.rs` and `write_ops.rs` next to the
struct in `mod.rs`.

Two things did not survive as pure code motion:

- The file-path and metadata readers (`config_file`, `status_file`,
  `load_config`, `read_status_len`) were associated fns on
  `ChunkedVectorsRead` that the writer reached through the type. They are
  now free `pub(super)` fns in `config.rs`, next to the types they read,
  so both sides get at them without widening visibility across the
  read-only boundary. `ChunkedVectorsRead::status_file` was `pub` but
  unused outside the module.

- `preopen_chunks` moved from `chunks.rs` to `read_only/lifecycle.rs`
  beside its only caller; `chunks.rs` keeps the shared chunk-name
  matching and gains a `chunks_prefix` helper for the two listing sites.

`ChunkedVectorsRead` is no longer re-exported from the component root, so
importers spell out `chunked_vectors::read_only::ChunkedVectorsRead` —
same as the sibling `dense` / `multi_dense` / `sparse` / `turbo`
read-only modules. No behaviour change.
