+++ lib/segment/src/vector_storage/chunked_vectors/chunks.rs
@@ -3,15 +3,21 @@ use std::path::{Path, PathBuf};
 use ahash::AHashMap;
 use common::mmap::{AdviceSetting, MULTI_MMAP_IS_SUPPORTED, create_and_ensure_length};
 use common::universal_io::{
-    CachedReadFs, OpenOptions, Populate, TypedStorage, UniversalIoError, UniversalRead,
-    UniversalReadFs, UniversalWrite,
+    OpenOptions, Populate, TypedStorage, UniversalIoError, UniversalRead, UniversalReadFs,
+    UniversalWrite,
 };
 
 use super::config::{MMAP_CHUNKS_PATTERN_END, MMAP_CHUNKS_PATTERN_START};
 
+/// Path prefix every chunk file shares. Listing by it keeps unrelated files in
+/// the directory from being enumerated at all.
+pub(super) fn chunks_prefix(directory: &Path) -> PathBuf {
+    directory.join(MMAP_CHUNKS_PATTERN_START)
+}
+
 /// Checks if the file name matches the pattern for mmap chunks
 /// Return ID from the file name if it matches, None otherwise
-fn check_mmap_file_name_pattern(file_name: &str) -> Option<usize> {
+pub(super) fn check_mmap_file_name_pattern(file_name: &str) -> Option<usize> {
     file_name
         .strip_prefix(MMAP_CHUNKS_PATTERN_START)
         .and_then(|file_name| file_name.strip_suffix(MMAP_CHUNKS_PATTERN_END))
@@ -29,33 +35,6 @@ pub fn chunk_open_options(
         populate,
         advice,
     }
-}
-
-/// Schedule background prefetch of every chunk file [`read_chunks`] will open.
-pub fn preopen_chunks(
-    fs: &impl CachedReadFs,
-    directory: &Path,
-    advice: AdviceSetting,
-    populate: Populate,
-) -> Result<(), UniversalIoError> {
-    let chunks_prefix = directory.join(MMAP_CHUNKS_PATTERN_START);
-    for listed in fs.list_files(&chunks_prefix)? {
-        let is_chunk = listed
-            .path
-            .file_name()
-            .and_then(|file_name| file_name.to_str())
-            .and_then(check_mmap_file_name_pattern)
-            .is_some();
-
-        if is_chunk {
-            fs.schedule_prefetch(
-                &listed.path,
-                Some(chunk_open_options(advice, populate, false)),
-                None,
-            )?;
-        }
-    }
-    Ok(())
 }
 
 pub fn read_chunks<T: bytemuck::Pod + Send, S: UniversalRead>(
@@ -77,11 +56,8 @@ pub fn read_chunks_from<T: bytemuck::Pod + Send, S: UniversalRead>(
     populate: Populate,
     writeable: bool,
 ) -> Result<Vec<TypedStorage<S, T>>, UniversalIoError> {
-    // List only the chunk files via the prefix, so unrelated files in the
-    // directory are never enumerated.
-    let chunks_prefix = directory.join(MMAP_CHUNKS_PATTERN_START);
     let mut chunks_files: AHashMap<usize, _> = AHashMap::new();
-    for listed in fs.list_files(&chunks_prefix)? {
+    for listed in fs.list_files(&chunks_prefix(directory))? {
         let path = listed.path;
         let chunk_id = path
             .file_name()
