+++ lib/segment/src/vector_storage/dense/read_only/lifecycle.rs
@@ -8,7 +8,7 @@ use crate::common::flags::in_memory_bitvec_flags::InMemoryBitvecFlags;
 use crate::common::operation_error::OperationResult;
 use crate::data_types::primitive::PrimitiveVectorElement;
 use crate::types::Distance;
-use crate::vector_storage::chunked_vectors::ChunkedVectorsRead;
+use crate::vector_storage::chunked_vectors::read_only::ChunkedVectorsRead;
 use crate::vector_storage::dense::appendable_dense_vector_storage::{
     DELETED_DIR_PATH, VECTORS_DIR_PATH,
 };
