+++ lib/segment/src/vector_storage/multi_dense/read_only/mod.rs
@@ -7,7 +7,7 @@ use crate::common::flags::in_memory_bitvec_flags::InMemoryBitvecFlags;
 use crate::common::operation_error::OperationResult;
 use crate::data_types::primitive::PrimitiveVectorElement;
 use crate::types::{Distance, MultiVectorConfig};
-use crate::vector_storage::chunked_vectors::ChunkedVectorsRead;
+use crate::vector_storage::chunked_vectors::read_only::ChunkedVectorsRead;
 
 mod lifecycle;
 mod live_reload;
